bail out if submodules are not initialized

+++ build.sh
@@ -5,6 +5,12 @@
 (
 cd $(dirname "${BASH_SOURCE[0]}")
 cd deps/luajit
+if [[ ! -e Makefile ]]
+then
+	echo "ERROR: LuaJIT submodule not initialized"
+	echo "Please run git submodule update --init"
+	exit 1
+fi
 make -j 8 'CFLAGS=-DLUAJIT_NUMMODE=2 -DLUAJIT_ENABLE_LUA52COMPAT'
 make install DESTDIR=$(pwd)
 cd ../dpdk
